Consolidate packet handling into shared proto file
Very big change to way packets are handled entirely.

Client & Server are more of a stub for proto crate, which will now act as a sort of shared function library. In a sense this will move all packet function code into a single file.

Server & Client packet code will instead be handled in 1 single .rs file, making it easier to develop and iterate easier.

I am not entirely sure if this will work as intended however optics are good so far.